requirements.txt
alabaster==0.7.12
attrs==21.2.0
Babel==2.9.1
bleach==4.1.0
certifi==2021.10.8
charset-normalizer==2.0.7
click==7.1.2
colorama==0.4.4
docopt==0.6.2
docutils==0.17
GH-IO-stubs>=7.7.21140
GH-Util-stubs>=7.7.21140
Grasshopper-stubs>=7.7.21140
honeybee-core>=1.47.26
honeybee-energy>=1.83.4
honeybee-schema>=1.47.0
honeybee-standards>=2.0.5
idna==3.3
imagesize==1.3.0
iniconfig==1.1.1
Jinja2==3.0.3
keyring==23.2.1
ladybug-core>=0.39.24
ladybug-geometry>=1.23.19
ladybug-geometry-polyskel>=1.3.40
ladybug-rhino>=1.32.0
MarkupSafe==2.0.1
packaging==21.2
pkginfo==1.7.1
pluggy==0.13.1
py==1.11.0
pydantic==1.8.2
pydantic-openapi-helper==0.2.7
Pygments==2.10.0
pyparsing==2.4.7
pytz==2021.3
readme-renderer==30.0
requests==2.26.0
requests-toolbelt==0.9.1
rfc3986==1.5.0
Rhino-stubs>=7.7.21140
six==1.16.0
snowballstemmer==2.1.0
toml==0.10.2
tqdm==4.62.3
typing-extensions==3.10.0.2
urllib3==1.26.7
webencodings==0.5.1
zipp==3.6.0
